#313: Options dialog has now been embedded into Launchpad.
