+++ Src/Orbiter/LpadTab.h
@@ -11,12 +11,13 @@
 // Property page indices
 #define PG_SCN  0
 #define PG_OPT  1
-#define PG_MOD  2
-#define PG_VID  3
-#define PG_JOY  4
-#define PG_EXT  5
-#define PG_ABT  6
-#define PG_WAIT 7
+#define PG_OPT2 2
+#define PG_MOD  3
+#define PG_VID  4
+#define PG_JOY  5
+#define PG_EXT  6
+#define PG_ABT  7
+#define PG_WAIT 8
 
 namespace orbiter {
 
@@ -41,8 +42,6 @@ namespace orbiter {
 		virtual void SetConfig(Config* cfg) {}
 		// Write config parameters back
 
-		virtual BOOL InitDialog(HWND hWnd, WPARAM wParam, LPARAM lParam) { return FALSE; }
-
 		virtual bool OpenHelp() { return false; }
 
 		/**
@@ -63,9 +62,6 @@ namespace orbiter {
 
 		void OpenTabHelp(const char* topic);
 
-		virtual BOOL Size(int w, int h);
-		// by default, this re-centers the items if RegisterItemPositions has been called
-
 		virtual void Show();
 		virtual void Hide();
 		virtual void LaunchpadShowing(bool show) {}
@@ -74,15 +70,20 @@ namespace orbiter {
 		inline HWND LaunchpadWnd() const { return pLp->hDlg; }
 		inline HINSTANCE AppInstance() const { return pLp->hInst; }
 
+		virtual BOOL OnInitDialog(HWND hWnd, WPARAM wParam, LPARAM lParam) { return FALSE; }
+
+		virtual BOOL OnSize(int w, int h);
+		// by default, this re-centers the items if RegisterItemPositions has been called
+
+		virtual BOOL OnNotify(HWND hDlg, int idCtrl, LPNMHDR pnmh) { return OnMessage(hDlg, WM_NOTIFY, (WPARAM)idCtrl, (LPARAM)pnmh); }
+
+		virtual BOOL OnMessage(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam) { return FALSE; }
+
 		virtual INT_PTR TabProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 		// generic message handler
 
 	protected:
 		HWND CreateTab(int resid);
-
-		//void RegisterItemPositions(int* _item, int _nitem);
-		// Keep a record of the positions of dialog items
-		// (for auto-recentering)
 
 		static INT_PTR CALLBACK TabProcHook(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
