+++ Src/Orbiter/Orbiter.rc
@@ -464,16 +464,17 @@ BEGIN
     PUSHBUTTON      "&Help",9,LAUNCHPAD_WIN_WIDTH-106,305,48,18
     PUSHBUTTON      "E&xit",IDEXIT,LAUNCHPAD_WIN_WIDTH-54,305,48,18
     PUSHBUTTON      "Scenarios ",IDC_MNU_SCN,6,43,59,14,BS_RIGHT | WS_GROUP
-    PUSHBUTTON      "Settings ",IDC_MNU_PRM,6,58,59,14,BS_RIGHT
-    PUSHBUTTON      "Modules ",IDC_MNU_MOD,6,73,59,14,BS_RIGHT
-    PUSHBUTTON      "Video ",IDC_MNU_VID,6,88,59,14,BS_RIGHT
-    PUSHBUTTON      "Joystick ",IDC_MNU_JOY,6,103,59,14,BS_RIGHT
-    PUSHBUTTON      "Extra ",IDC_MNU_EXT,6,118,59,14,BS_RIGHT
-    PUSHBUTTON      "About Orbiter ",IDC_MNU_ABT,6,133,59,14,BS_RIGHT
+	PUSHBUTTON      "Options ",IDC_MNU_OPT,6,58,59,14,BS_RIGHT
+    PUSHBUTTON      "Settings ",IDC_MNU_PRM,6,73,59,14,BS_RIGHT
+    PUSHBUTTON      "Modules ",IDC_MNU_MOD,6,88,59,14,BS_RIGHT
+    PUSHBUTTON      "Video ",IDC_MNU_VID,6,103,59,14,BS_RIGHT
+    PUSHBUTTON      "Joystick ",IDC_MNU_JOY,6,118,59,14,BS_RIGHT
+    PUSHBUTTON      "Extra ",IDC_MNU_EXT,6,133,59,14,BS_RIGHT
+    PUSHBUTTON      "About Orbiter ",IDC_MNU_ABT,6,148,59,14,BS_RIGHT
     CTEXT           "\r\norbit.medphys.ucl.ac.uk\r\n(c) 2000-2016\r\nMartin Schweiger",IDC_BLACKBOX,248,0,152,37,0,WS_EX_RIGHT
     CONTROL         IDB_BANNER,IDC_LOGO,"Static",SS_BITMAP,0,0,248,37
     CONTROL         "",IDC_SHADOW,"Static",SS_OWNERDRAW,0,37,5,5
-	CONTROL         "",IDC_MNU_PAGECONTAINER,"STATIC",SS_OWNERDRAW | WS_CLIPCHILDREN,69,43,LAUNCHPAD_WIN_WIDTH-75,256,WS_EX_CONTROLPARENT
+	CONTROL         "",IDC_MNU_PAGECONTAINER,"STATIC",SS_OWNERDRAW | WS_CLIPCHILDREN,69,43,LAUNCHPAD_WIN_WIDTH-75,256,WS_EX_NOACTIVATE | WS_EX_CONTROLPARENT
     CTEXT           "v.123456",IDC_VERSION,4,315,44,8,WS_DISABLED
 END
 
@@ -490,6 +491,15 @@ BEGIN
     CONTROL         "Custom2",IDC_SCN_SPLIT1,"OrbiterDlgCtrl",WS_CLIPSIBLINGS,0,12,273,221
     PUSHBUTTON      "Save current ...",IDC_SCN_SAVE,0,237,74,14
     PUSHBUTTON      "Clear quicksaves",IDC_SCN_DELQS,76,237,74,14
+END
+
+IDD_PAGE_OPT DIALOGEX 0, 0, 277, 254
+STYLE DS_SETFONT | DS_CONTROL | WS_CHILD
+FONT 8, "MS Shell Dlg", 400, 0, 0x0
+BEGIN
+    CONTROL         "",IDC_OPT_SPLIT,"OrbiterDlgCtrl",WS_CLIPSIBLINGS | WS_CLIPCHILDREN,0,0,277,254
+	CONTROL         "Tree1",IDC_OPT_PAGELIST,"SysTreeView32",TVS_SHOWSELALWAYS | WS_BORDER | WS_HSCROLL | WS_TABSTOP,0,0,120,254
+	CONTROL         "",IDC_OPT_PAGECONTAINER,"OrbiterDlgCtrl",WS_CLIPSIBLINGS,130,0,147,254,WS_EX_NOACTIVATE | WS_EX_CONTROLPARENT
 END
 
 IDD_PAGE_VIS DIALOGEX 0, 0, 277, 254
@@ -698,7 +708,7 @@ IDD_OPTIONS_CELSPHERE DIALOGEX 0, 0, 220, 146
 STYLE DS_SETFONT | DS_CONTROL | WS_CHILD
 FONT 8, "MS Shell Dlg", 0, 0, 0x0
 BEGIN
-    CONTROL         "Background stars: show as pixels",IDC_OPT_CSP_ENABLESTARPIX,"Button",BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP,6,0,160,10
+	CONTROL         "Background stars: show as pixels",IDC_OPT_CSP_ENABLESTARPIX,"Button",BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP,6,0,160,10
 	LTEXT           "Map app. magnitude",IDC_STATIC2,16,16,68,8
 	LTEXT           "to max. brightness 1",IDC_STATIC3,121,16,64,8
 	EDITTEXT        IDC_OPT_CSP_STARMAGHI,84,14,22,14,ES_AUTOHSCROLL | ES_READONLY
@@ -709,15 +719,15 @@ BEGIN
 	CONTROL         "",IDC_OPT_CSP_STARMAGLOSPIN,"msctls_updown32",UDS_ARROWKEYS,106,30,10,14
 	EDITTEXT        IDC_OPT_CSP_STARMINBRT,180,30,22,14,ES_AUTOHSCROLL | ES_READONLY
 	CONTROL         "",IDC_OPT_CSP_STARMINBRTSPIN,"msctls_updown32",UDS_ARROWKEYS,202,30,10,14
-    LTEXT           "Magnitude-brightness mapping:",IDC_STATIC6,16,48,99,8
-    CONTROL         "linear",IDC_OPT_CSP_STARMAPLIN,"Button",BS_AUTORADIOBUTTON,127,48,33,10
-    CONTROL         "exponential",IDC_OPT_CSP_STARMAPEXP,"Button",BS_AUTORADIOBUTTON,165,48,52,10
-    CONTROL         "Background stars: show as image",IDC_OPT_CSP_ENABLESTARMAP,"Button",BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP,6,65,160,10
-    COMBOBOX        IDC_OPT_CSP_STARMAPIMAGE,16,79,170,49,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
+	LTEXT           "Magnitude-brightness mapping:",IDC_STATIC6,16,48,99,8
+	CONTROL         "linear",IDC_OPT_CSP_STARMAPLIN,"Button",BS_AUTORADIOBUTTON,127,48,33,10
+	CONTROL         "exponential",IDC_OPT_CSP_STARMAPEXP,"Button",BS_AUTORADIOBUTTON,165,48,52,10
+	CONTROL         "Background stars: show as image",IDC_OPT_CSP_ENABLESTARMAP,"Button",BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP,6,65,160,10
+	COMBOBOX        IDC_OPT_CSP_STARMAPIMAGE,16,79,170,49,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
 	CONTROL         "Background map",IDC_OPT_CSP_ENABLEBKGMAP,"Button",BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP,6,104,160,10
-    COMBOBOX        IDC_OPT_CSP_BKGIMAGE,16,118,170,49,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
+	COMBOBOX        IDC_OPT_CSP_BKGIMAGE,16,118,170,49,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
 	LTEXT           "Brightness:",IDC_STATIC1,16,137,40,8
-    CONTROL         "",IDC_OPT_CSP_BGBRIGHTNESS,"OrbiterCtrl_Gauge",WS_TABSTOP,56,136,130,10
+	CONTROL         "",IDC_OPT_CSP_BGBRIGHTNESS,"OrbiterCtrl_Gauge",WS_TABSTOP,56,136,130,10
 END
 
 IDD_OPTIONS_VISHELPER DIALOGEX 0, 0, 220, 122
