+++ Src/Orbiter/TabScenario.h
@@ -9,6 +9,7 @@
 #ifndef __TABSCENARIO_H
 #define __TABSCENARIO_H
 
+#include <CommCtrl.h>
 #include "LpadTab.h"
 #include "CustomControls.h"
 
@@ -28,7 +29,7 @@ namespace orbiter {
 
 		bool DynamicSize() const { return true; }
 
-		BOOL Size(int w, int h);
+		BOOL OnSize(int w, int h);
 
 		int GetSelScenario(char* scn, int len);
 		// returns name of currently selected scenario file
@@ -36,7 +37,9 @@ namespace orbiter {
 
 		void LaunchpadShowing(bool show);
 
-		INT_PTR TabProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
+		BOOL OnNotify(HWND hDlg, int idCtrl, LPNMHDR pnmh);
+
+		BOOL OnMessage(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 
 	protected:
 		void RefreshList(bool preserveSelection);
